# openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce Backend API
  description: API documentation for the E-Commerce backend. Can be imported into Postman.
  version: 1.0.0
servers:
  - url: http://localhost:9090
    description: Local development server
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User registered successfully
        '400':
          description: Username or email already exists
  /api/auth/login:
    post:
      summary: Authenticate user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Unauthorized
  /api/products:
    get:
      summary: Get all products
      tags:
        - Products
      responses:
        '200':
          description: A list of products
  /api/cart:
    get:
      summary: Get user cart
      tags:
        - Cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's cart details
        '401':
          description: Unauthorized
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT